pyproject.toml
[project]
name = "Agent-Evals"
version = "2.0.0"
description = "An integration of Pytest Full Stack, Evidently and AI Agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=4.0.0",
    "bs4>=0.0.2",
    "chromadb>=1.0.15",
    "dataclasses>=0.8",
    "ddgs>=9.0.0",
    "enum34>=1.1.10",
    "evaluate>=0.4.4",
    "evidently[llm]>=0.7.9",
    "faiss-cpu>=1.11.0.post1",
    "filetype>=1.2.0",
    "google-api-python-client>=2.177.0",
    "gradio>=5.35.0",
    "groq>=0.30.0",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "jupyter>=1.1.1",
    "langchain>=0.2.0,<0.3.0",
    "langchain-community>=0.2.0,<0.3.0",
    "langchain-experimental>=0.0.60,<0.1.0",
    "langchain-groq>=0.1.0,<0.2.0",
    "langchain-openai>=0.1.0,<0.2.0",
    "langgraph>=0.2.0,<0.3.0",
    "langgraph-checkpoint-sqlite>=1.0.0,<2.0.0",
    "libmagic>=1.0",
    "linkup-sdk>=0.2.8",
    "matplotlib>=3.10.5",
    "mistralai>=1.9.3",
    "mock>=5.2.0",
    "nltk>=3.9.1",
    "numpy>=1.26.4",
    "openai>=1.93.0",
    "openai-agents>=0.1.0",
    "pandas>=2.3.0",
    "pillow>=11.3.0",
    "playwright>=1.53.0",
    "pyboxen>=1.3.0",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-bdd>=8.1.0",
    "pytest-cov>=6.2.1",
    "pytest-html>=4.1.1",
    "pytest-playwright>=0.7.0",
    "python-dotenv>=1.1.1",
    "python-fasthtml>=0.12.25",
    "python-magic-bin>=0.4.14",
    "python-multipart>=0.0.20",
    "ragas>=0.2.15",
    "requests>=2.32.4",
    "rich>=14.0.0",
    "rouge>=1.0.1",
    "scikit-learn>=1.7.0",
    "scipy>=1.16.0",
    "seaborn>=0.13.2",
    "sendgrid>=6.12.4",
    "sqlmodel>=0.0.24",
    "tensorflow>=2.18.1",
    "tracely>=0.2.5",
    "twitter>=1.19.6",
    "unstructured[md]>=0.18.2",
    "wikipedia>=1.4.0",
]
